🐛 Fix bug in CSES 1083 solution
The max value for the nth partial sum of n is (2 * 10^5) * (2 * 10^5 + 1) / 2; which is larger than the max size of an int. Thus, we need to use a long long instead; using ll for n as well to avoid nasty casting in the formula.😊

+++ cses/missing_number.cpp
@@ -61,12 +61,12 @@ typedef long long ll;
 // O(n)
 int main()
 {
-  int n;
+  ll n;
   cin >> n;
 
   // O(1) 😂
   // the nth partial sum
-  int solution = n * (n + 1) / 2;
+  ll solution = n * (n + 1) / 2;
 
   // O(n) loop (read in the input)
   while (--n)
